Create packages for Password Manager
